Rewrite into F#, Fable & Elmish
Elm 0.19.1 is quite good, but it kinda unfit for my
needs (lack of well maintained foundational libraries such
as auto encoder/decoder of JSON and other things that comes
for free in Fable ecosystem).

I was going to write the project in F# first anyway, but
quite curious with Elm and build setup minimalism.